+++ packages/react-widgets/src/types.d.ts
@@ -1,4 +1,5 @@
 import { AggregationTypes } from '@carto/react-core';
+import type { SxProps, Theme } from '@mui/material';
 
 type CommonWidgetProps = {
   id: string,
@@ -41,6 +42,8 @@ export type BarWidget = {
 export type FormulaWidget = CommonWidgetProps & MonoColumnWidgetProps;
 
 export type GeocoderWidget = {
+  className: string,
+  sx?: SxProps<Theme>,
   onError?: Function
 }
 
@@ -94,6 +97,7 @@ export type TimeSeriesWidget = {
 } & CommonWidgetProps & MonoColumnWidgetProps;
 
 export type LegendWidget = {
+  className?: string;
   initialCollapsed?: boolean;
   customLegendTypes?: Record<string, Function>;
   layerOrder?: string[];
@@ -108,6 +112,7 @@ export type WidgetWithAlert = {
 }
 
 export type FeatureSelectionWidget = {
+  className?: string;
   selectionModes?: string[],
   editModes?: string[],
   tooltipPlacement?: string,