+++ packages/react-widgets/src/widgets/FeatureSelectionWidget.js
@@ -43,6 +43,7 @@ const EDIT_MODES_MAP = {
 };
 
 function FeatureSelectionWidget({
+  className,
   selectionModes: selectionModesKeys,
   editModes: editModesKeys,
   tooltipPlacement,
@@ -96,6 +97,7 @@ function FeatureSelectionWidget({
 
   return (
     <FeatureSelectionWidgetUI
+      className={className}
       selectionModes={selectionModes}
       editModes={editModes}
       selectedMode={selectedMode}
@@ -120,6 +122,7 @@ FeatureSelectionWidget.defaultProps = {
 };
 
 FeatureSelectionWidget.propTypes = {
+  className: FeatureSelectionWidgetUI.propTypes.className,
   selectionModes: PropTypes.arrayOf(
     PropTypes.oneOf(Object.values(FEATURE_SELECTION_MODES))
   ),