+++ packages/react-widgets/src/widgets/LegendWidget.js
@@ -9,12 +9,14 @@ import sortLayers from './utils/sortLayers';
  * Renders a <LegendWidget /> component
  * @param  {object} props
  * @param  {string} [props.title] - Title of the widget.
+ * @param  {string} [props.className] - CSS class name.
  * @param  {Object.<string, Function>} [props.customLayerOptions] - Allow to add custom controls to a legend item to tweak the associated layer.
  * @param  {Object.<string, Function>} [props.customLegendTypes] - Allow to customise by default legend types that can be rendered.
  * @param  {boolean} [props.initialCollapsed] - Define initial collapsed value. false by default.
  * @param  {string[]} [props.layerOrder] - Array of layer identifiers. Defines the order of layer legends. [] by default.
  */
 function LegendWidget({
+  className,
   customLayerOptions,
   customLegendTypes,
   initialCollapsed,
@@ -64,6 +66,7 @@ function LegendWidget({
   return (
     <LegendWidgetUI
       title={title}
+      className={className}
       customLegendTypes={customLegendTypes}
       customLayerOptions={customLayerOptions}
       layers={layers}
@@ -78,6 +81,7 @@ function LegendWidget({
 
 LegendWidget.propTypes = {
   title: PropTypes.string,
+  className: PropTypes.string,
   customLegendTypes: PropTypes.objectOf(PropTypes.func),
   initialCollapsed: PropTypes.bool,
   layerOrder: PropTypes.arrayOf(PropTypes.string)